refactor: rename overall dataset file and update references throughout the codebase
